backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
redis
python-dotenv
dj-database-url
